[smaf] update for a file conatians mmmg related chunks

+++ src/main/java/vavi/sound/smaf/chunk/MasterTrackSequenceDataChunk.java
@@ -37,6 +37,6 @@ logger.log(Level.DEBUG, "MasterTrackSequenceData: " + size + " bytes");
     @Override
     protected void init(CrcDataInputStream dis, Chunk parent)
         throws InvalidSmafDataException, IOException {
-dis.skipBytes((int) (long) size); // TODO
+dis.skipBytes(size); // TODO
     }
 }
